Move code to extrinsic_types
