+++ subxt/src/blocks/mod.rs
@@ -6,6 +6,8 @@
 
 mod block_types;
 mod blocks_client;
+mod extrinsic_types;
 
-pub use block_types::{Block, ExtrinsicDetails, ExtrinsicEvents, RootExtrinsic, StaticExtrinsic};
+pub use block_types::Block;
 pub use blocks_client::{subscribe_to_block_headers_filling_in_gaps, BlocksClient};
+pub use extrinsic_types::{ExtrinsicDetails, ExtrinsicEvents, RootExtrinsic, StaticExtrinsic};
